% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{available_nar_versions}
\alias{available_nar_versions}
\title{Scrape availabe NAR versions from the StatCan website}
\usage{
available_nar_versions(refresh = FALSE)
}
\arguments{
\item{refresh}{Logical indicating whether to refresh the cached version list}
}
\value{
A tibble with available NAR versions and their URLs
}
\description{
Scrape availabe NAR versions from the StatCan website
}
\examples{
\dontrun{
versions <- available_nar_versions()
}
}
