% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{collect_sf}
\alias{collect_sf}
\title{Collect sf object from a remote table}
\usage{
collect_sf(tbl)
}
\arguments{
\item{tbl}{Remote table to collect}
}
\value{
An sf object
}
\description{
Collect sf object from a remote table
}
\examples{
con <- nar_connection()
nar_sf <- con |>
  head(20) |>
  collect_sf()
}
