% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{nar_cached_versions}
\alias{nar_cached_versions}
\title{Versions already present in the local cache}
\usage{
nar_cached_versions(cache_path)
}
\arguments{
\item{cache_path}{Directory holding the `<version>.duckdb` files}
}
\value{
Character vector of version keys, newest first
}
\description{
Versions already present in the local cache
}
\keyword{internal}
