% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_helpers.R
\name{nar_con}
\alias{nar_con}
\title{Extract the DBI connection behind a lazy table}
\usage{
nar_con(x)
}
\arguments{
\item{x}{A lazy table or a DBI connection}
}
\value{
A DBI connection, or NULL
}
\description{
Extract the DBI connection behind a lazy table
}
\keyword{internal}
