% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{nar_covers_set}
\alias{nar_covers_set}
\title{Does one coverage set satisfy another?}
\usage{
nar_covers_set(have, want)
}
\arguments{
\item{have}{`"ALL"` or a character vector of abbreviations}

\item{want}{`"ALL"` or a character vector of abbreviations}
}
\value{
`TRUE` when nothing further needs downloading
}
\description{
Does one coverage set satisfy another?
}
\keyword{internal}
