% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_lexicon.R
\name{nar_is_street_type}
\alias{nar_is_street_type}
\title{Is this token a street type in the given language context?}
\usage{
nar_is_street_type(surface, lang = "en")
}
\arguments{
\item{surface}{A character vector of already-folded surface forms}

\item{lang}{A character vector of `"en"` / `"fr"`, recycled against `surface`}
}
\value{
A logical vector
}
\description{
Is this token a street type in the given language context?
}
\keyword{internal}
