% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known.R
\name{nar_known_keys}
\alias{nar_known_keys}
\title{The components `known` accepts}
\usage{
nar_known_keys()
}
\value{
A character vector of column names
}
\description{
The components `known` accepts
}
\keyword{internal}
