% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar_zip.R
\name{nar_le_raw}
\alias{nar_le_raw}
\title{Encode a number as little-endian bytes}
\usage{
nar_le_raw(x, n)
}
\arguments{
\item{x}{A non-negative number}

\item{n}{Field width in bytes}
}
\value{
A raw vector of length `n`
}
\description{
Encode a number as little-endian bytes
}
\keyword{internal}
