% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{nar_location_types}
\alias{nar_location_types}
\title{Non-string column types in the NAR location file}
\usage{
nar_location_types()
}
\value{
A named list of `arrow` types
}
\description{
Non-string column types in the NAR location file
}
\keyword{internal}
