% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_address.R
\name{nar_normalized_columns}
\alias{nar_normalized_columns}
\title{The NAR-shaped columns normalize_address() produces}
\usage{
nar_normalized_columns()
}
\value{
A character vector of column names
}
\description{
The NAR-shaped columns normalize_address() produces
}
\keyword{internal}
