% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_address.R
\name{nar_split_commas}
\alias{nar_split_commas}
\title{Split a token vector on comma tokens}
\usage{
nar_split_commas(toks)
}
\arguments{
\item{toks}{A character vector of tokens}
}
\value{
A list of non-empty token vectors
}
\description{
Split a token vector on comma tokens
}
\keyword{internal}
