% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_helpers.R
\name{nar_storage_crs}
\alias{nar_storage_crs}
\title{CRS in which NAR geometry is stored}
\usage{
nar_storage_crs()
}
\value{
CRS identifier string
}
\description{
CRS in which NAR geometry is stored
}
\keyword{internal}
