% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_address.R
\name{nar_take_trailing_unit}
\alias{nar_take_trailing_unit}
\title{Take a trailing unit designator off the end of a street}
\usage{
nar_take_trailing_unit(toks)
}
\arguments{
\item{toks}{A character vector of tokens}
}
\value{
A list with `unit` and `rest`
}
\description{
Take a trailing unit designator off the end of a street
}
\keyword{internal}
