% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_address.R
\name{nar_tokens}
\alias{nar_tokens}
\title{Split a normalized string into tokens, commas included}
\usage{
nar_tokens(s)
}
\arguments{
\item{s}{A single normalized address string}
}
\value{
A character vector of tokens
}
\description{
Split a normalized string into tokens, commas included
}
\keyword{internal}
