% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{nar_unit_sql}
\alias{nar_unit_sql}
\title{The stored side of the unit comparison}
\usage{
nar_unit_sql(col)
}
\arguments{
\item{col}{The stored unit column, qualified}
}
\value{
A SQL fragment
}
\description{
Case, spaces and periods only -- 592,561 of NAR's units are
not upper case and 16 are untrimmed. The vocabulary translation happens on
the input, in [nar_unit_fold()].
}
\keyword{internal}
