% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{normalized_nar_version}
\alias{normalized_nar_version}
\title{Normalize NAR version string}
\usage{
normalized_nar_version(version, refresh = FALSE)
}
\arguments{
\item{version}{Version of the NAR database to connect to. Default is "latest}

\item{refresh}{Logical indicating whether to refresh the cached version list}
}
\value{
Normalized version string
}
\description{
Normalize NAR version string
}
\keyword{internal}
