% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{collect_nar}
\alias{collect_nar}
\title{Collect sf object from a car connection}
\usage{
collect_nar(tbl)
}
\arguments{
\item{tbl}{nar table to collect}
}
\value{
An sf object
}
\description{
Collect sf object from a car connection
}
\examples{
con <- nar_connection()
nar_sf <- con |>
  head(20) |>
  collect_nar()
}
