% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{nar_connection}
\alias{nar_connection}
\title{Get NAR data}
\usage{
nar_connection(
  version = "latest",
  format = c("duckdb", "parquet"),
  refresh = FALSE
)
}
\arguments{
\item{version}{Version of the NAR database to connect to. Default is "latest".}

\item{format}{Format of the database. Currently only "duckdb" and "parquet" are supported.}

\item{refresh}{Logical indicating whether to refresh the local cache of the NAR database.}
}
\value{
A connection to the NAR database
}
\description{
This function downloads the NAR data if necessary and returns a connection the NAR database
}
\examples{
con <- nar_connection()
}
