% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{nar_connection}
\alias{nar_connection}
\title{Get NAR data}
\usage{
nar_connection(version = "latest", refresh = FALSE)
}
\arguments{
\item{version}{Version of the NAR database to connect to. Default is "latest".}

\item{refresh}{Logical indicating whether to refresh the local cache of the NAR database.}
}
\value{
A connection to the NAR database containing Addresses and Locations tables
}
\description{
This function downloads the NAR data if necessary and returns a connection the NAR database
}
\examples{
con <- nar_connection()
}
