% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse Geocode Coordinates to Address}
\usage{
reverse_geocode(
  x,
  match_radius = 100,
  output = "multiple",
  source = "nar",
  ...
)
}
\arguments{
\item{x}{An `sf` POINT object with coordinates to reverse geocode}

\item{match_radius}{Maximum distance (in meters) to search for matching addresses (default
is 100 meters)}

\item{output}{Type of output to return. Options are "address" (returns a single formatted address string),
"components" (returns a data frame with address components for the closest match), or "multiple" (returns a data frame with all matches within the match radius).
Default is "multiple".}

\item{source}{Source dataset to use for reverse geocoding. Currently only "nar" (National Address Repository) is supported.}

\item{...}{Additional arguments (currently unused)}
}
\value{
Depending on the `output` parameter, either a single address string, a data frame with address components for the closest match, or a data frame with all matches within the match radius.
}
\description{
Determines closest address(s) to given coordinates using the NAR dataset based on maximym match radius
}
\examples{
\dontrun{
library(sf)
# Create an sf POINT object with coordinates to reverse geocode
point <- st_sfc(st_point(c(-75.6972, 45.      4215)), crs = 4326)
# Reverse geocode the point to get the closest address
address <- reverse_geocode(point, match_radius = 200, output = "address")
print(address)
}
}
